test_python/requirements.txt
fastapi
uvicorn
pydantic
bcrypt
PyJWT
pytest
python-multipart
